directory: move Modals in directory
